requirements.txt
# Create requirements.txt
echo "Django==4.2.7
psycopg2-binary==2.9.9
django-crispy-forms==2.1
django-crispy-bootstrap5==2023.10
django-allauth==0.57.0
django-guardian==2.4.0
django-filter==23.5
django-extensions==3.2.3
python-dotenv==1.0.0
Pillow==10.1.0
reportlab==4.0.8
openpyxl==3.1.2
django-import-export==3.3.4
django-debug-toolbar==4.2.0